Created catalog_utils.py file to keep the low-level routines to help handling the catalogs.
